refactor(checker): split contextual-parameter utilities out of utilities/core.rs
Extract the contextual-parameter and callable-contextual type-resolution
methods from crates/tsz-checker/src/types/utilities/core.rs into a new sibling
submodule utilities/contextual_parameters.rs. Pure code move (one self-contained
cluster of CheckerState methods, no logic change): core.rs drops 2779 -> 1703
LOC, back under the 2000-LOC checker boundary, so it is removed from the
arch_guard LINE_LIMIT_CHECKS allowlist (9 -> 8 oversized checker files).

The moved cluster's private helpers are only called within the cluster
(verified), and no moved method forward-references a helper left in core.rs, so
no visibility changes were needed beyond keeping the existing pub(crate) entry
points. arch_guard passes; tsz-checker builds clean with no new warnings.

AgentName: M1-Opus
